requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LLM Integration (Google Gemini)
google-generativeai>=0.8.0
langchain==0.1.6
langchain-google-genai>=2.0.0
langchain-community==0.0.19

# Vector Database (ChromaDB)
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Date/Time Handling
python-dateutil==2.8.2
pytz==2024.1

# CORS & Security
python-multipart==0.0.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
